New API service for use in prosecution concluded listener & scheduler
