+++ crown-court-proceeding/src/main/java/uk/gov/justice/laa/crime/crowncourt/prosecution_concluded/impl/ProcessSentencingImpl.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import uk.gov.justice.laa.crime.crowncourt.model.UpdateSentenceOrder;
-import uk.gov.justice.laa.crime.crowncourt.service.MaatCourtDataService;
+import uk.gov.justice.laa.crime.crowncourt.prosecution_concluded.service.CourtDataAPIService;
 import uk.gov.justice.laa.crime.crowncourt.util.DateUtil;
 
 import java.time.LocalDate;
@@ -17,7 +17,7 @@ import static uk.gov.justice.laa.crime.crowncourt.staticdata.enums.CrownCourtCas
 @RequiredArgsConstructor
 public class ProcessSentencingImpl {
 
-    private final MaatCourtDataService maatCourtDataService;
+    private final CourtDataAPIService courtDataAPIService;
 
     @Value("${spring.datasource.username}")
     private String dbUser;
@@ -29,7 +29,7 @@ public class ProcessSentencingImpl {
         if (caseEndDate != null) {
             String user = dbUser != null ? dbUser.toUpperCase() : null;
             if (APPEAL_CC.getValue().equalsIgnoreCase(catyType)) {
-                maatCourtDataService
+                courtDataAPIService
                         .invokeUpdateAppealSentenceOrderDate(UpdateSentenceOrder.builder()
                                 .repId(maatId)
                                 .dbUser(user)
@@ -38,7 +38,7 @@ public class ProcessSentencingImpl {
                                 .build()
                         );
             } else {
-                maatCourtDataService
+                courtDataAPIService
                         .invokeUpdateSentenceOrderDate(UpdateSentenceOrder.builder()
                                 .repId(maatId)
                                 .dbUser(user)
