+++ crown-court-proceeding/src/test/java/uk/gov/justice/laa/crime/crowncourt/service/MaatCourtDataServiceTest.java
@@ -3,9 +3,10 @@ package uk.gov.justice.laa.crime.crowncourt.service;
 import org.assertj.core.api.junit.jupiter.SoftAssertionsExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.*;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import uk.gov.justice.laa.crime.commons.client.RestAPIClient;
@@ -13,19 +14,13 @@ import uk.gov.justice.laa.crime.commons.exception.APIClientException;
 import uk.gov.justice.laa.crime.crowncourt.config.MockServicesConfiguration;
 import uk.gov.justice.laa.crime.crowncourt.config.ServicesConfiguration;
 import uk.gov.justice.laa.crime.crowncourt.data.builder.TestModelDataBuilder;
-import uk.gov.justice.laa.crime.crowncourt.dto.maatcourtdata.*;
-import uk.gov.justice.laa.crime.crowncourt.model.UpdateCCOutcome;
-import uk.gov.justice.laa.crime.crowncourt.model.UpdateSentenceOrder;
-import uk.gov.justice.laa.crime.crowncourt.prosecution_concluded.model.ProsecutionConcluded;
+import uk.gov.justice.laa.crime.crowncourt.dto.maatcourtdata.IOJAppealDTO;
+import uk.gov.justice.laa.crime.crowncourt.dto.maatcourtdata.RepOrderCCOutcomeDTO;
+import uk.gov.justice.laa.crime.crowncourt.dto.maatcourtdata.UpdateRepOrderRequestDTO;
 
-import java.util.List;
-import java.util.UUID;
-
-import static java.util.Collections.emptyMap;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
@@ -50,7 +45,9 @@ class MaatCourtDataServiceTest {
                 .thenReturn(expected);
 
         IOJAppealDTO response =
-                maatCourtDataService.getCurrentPassedIOJAppealFromRepId(TestModelDataBuilder.TEST_REP_ID, LAA_TRANSACTION_ID);
+                maatCourtDataService.getCurrentPassedIOJAppealFromRepId(
+                        TestModelDataBuilder.TEST_REP_ID, LAA_TRANSACTION_ID
+                );
 
         assertThat(response).isEqualTo(expected);
     }
@@ -100,149 +97,6 @@ class MaatCourtDataServiceTest {
         assertThatThrownBy(() -> maatCourtDataService.getRepOrderCCOutcomeByRepId(
                 TestModelDataBuilder.TEST_REP_ID, LAA_TRANSACTION_ID)
         ).isInstanceOf(APIClientException.class);
-    }
-    @Test
-    void givenAInvalidParameter_whenRetrieveHearingForCaseConclusionIsInvoked_thenEmptyIsReturned() {
-        WQHearingDTO wqHearingDTO = maatCourtDataService.retrieveHearingForCaseConclusion(ProsecutionConcluded.builder().hearingIdWhereChangeOccurred(UUID.randomUUID())
-                .maatId(TestModelDataBuilder.TEST_REP_ID).build());
-        assertThat(wqHearingDTO).isNull();
-    }
-
-    @Test
-    void givenAInvalidParameter_whenFindWQLinkRegisterByMaatIdIsInvoked_thenNullIsReturned() {
-        when(maatAPIClient.get(any(), any(), any(), any())).thenReturn(null);
-        int caseId = maatCourtDataService.findWQLinkRegisterByMaatId(TestModelDataBuilder.TEST_REP_ID);
-        verify(maatAPIClient).get(any(), any(), any(), any());
-        assertThat(caseId).isZero();
-    }
-
-    @Test
-    void givenAInvalidParameter_whenFindWQLinkRegisterByMaatIdIsInvoked_thenEmptyIsReturned() {
-        when(maatAPIClient.get(any(), any(), any(), any())).thenReturn(List.of());
-        int caseId = maatCourtDataService.findWQLinkRegisterByMaatId(TestModelDataBuilder.TEST_REP_ID);
-        assertThat(caseId).isZero();
-    }
-
-    @Test
-    void givenAValidParameter_whenFindWQLinkRegisterByMaatIdIsInvoked_thenResponseIsReturned() {
-        when(maatAPIClient.get(any(), any(), any(), any())).thenReturn(List.of(WQLinkRegisterDTO.builder()
-                .caseId(TestModelDataBuilder.TEST_CASE_ID).build()));
-        int caseId = maatCourtDataService.findWQLinkRegisterByMaatId(TestModelDataBuilder.TEST_REP_ID);
-        assertThat(caseId).isEqualTo(TestModelDataBuilder.TEST_CASE_ID);
-    }
-
-    @Test
-    void givenAValidParameter_whenFindOffenceByCaseIdIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.findOffenceByCaseId(TestModelDataBuilder.TEST_REP_ID);
-        verify(maatAPIClient).get(any(), any(), any(), any());
-    }
-
-    @Test
-    void givenAInvalidParameter_whenGetOffenceNewOffenceCountIsInvoked_thenNullIsReturned() {
-        when(maatAPIClient.get(any(), anyString(), anyMap(), anyInt(), anyString())).thenReturn(null);
-        int offenceCount = maatCourtDataService.getOffenceNewOffenceCount(TestModelDataBuilder.TEST_CASE_ID, TestModelDataBuilder.TEST_OFFENCE_ID);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyString());
-        assertThat(offenceCount).isZero();
-    }
-
-    @Test
-    void givenAInvalidParameter_whenGetOffenceNewOffenceCountIsInvoked_thenEmptyIsReturned() {
-        when(maatAPIClient.get(any(), anyString(), anyMap(), anyInt(), anyString())).thenReturn(List.of());
-        int offenceCount = maatCourtDataService.getOffenceNewOffenceCount(TestModelDataBuilder.TEST_CASE_ID, TestModelDataBuilder.TEST_OFFENCE_ID);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyString());
-        assertThat(offenceCount).isZero();
-    }
-
-    @Test
-    void givenAValidParameter_whenGetOffenceNewOffenceCountIsInvoked_thenResponseIsReturned() {
-        when(maatAPIClient.get(any(), anyString(), anyMap(), anyInt(), anyString())).thenReturn(List.of(4));
-        int offenceCount = maatCourtDataService.getOffenceNewOffenceCount(TestModelDataBuilder.TEST_CASE_ID, TestModelDataBuilder.TEST_OFFENCE_ID);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyString());
-        assertThat(offenceCount).isEqualTo(4);
-    }
-
-    @Test
-    void givenAInvalidParameter_whenGetWQOffenceNewOffenceCountIsInvoked_thenNullIsReturned() {
-        when(maatAPIClient.get(any(), anyString(), anyMap(), anyInt(), anyString())).thenReturn(null);
-        int offenceCount = maatCourtDataService.getWQOffenceNewOffenceCount(TestModelDataBuilder.TEST_CASE_ID, TestModelDataBuilder.TEST_OFFENCE_ID);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyString());
-        assertThat(offenceCount).isZero();
-    }
-
-    @Test
-    void givenAInvalidParameter_whenGetWQOffenceNewOffenceCountIsInvoked_thenEmptyIsReturned() {
-        when(maatAPIClient.get(any(), anyString(), anyMap(), anyInt(), anyString())).thenReturn(List.of());
-        int offenceCount = maatCourtDataService.getWQOffenceNewOffenceCount(TestModelDataBuilder.TEST_CASE_ID, TestModelDataBuilder.TEST_OFFENCE_ID);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyString());
-        assertThat(offenceCount).isZero();
-    }
-
-    @Test
-    void givenAValidParameter_whenGetWQOffenceNewOffenceCountIsInvoked_thenResponseIsReturned() {
-        when(maatAPIClient.get(any(), anyString(), anyMap(), anyInt(), anyString())).thenReturn(List.of(5));
-        int offenceCount = maatCourtDataService.getWQOffenceNewOffenceCount(TestModelDataBuilder.TEST_CASE_ID, TestModelDataBuilder.TEST_OFFENCE_ID);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyString());
-        assertThat(offenceCount).isEqualTo(5);
-    }
-
-    @Test
-    void givenAValidParameter_whenFindResultsByWQTypeSubTypeIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.findResultsByWQTypeSubType(1244, 25243);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyInt());
-    }
-
-    @Test
-    void givenAValidParameter_whenGetResultCodeByCaseIdAndAsnSeqIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.getResultCodeByCaseIdAndAsnSeq(TestModelDataBuilder.TEST_CASE_ID, TestModelDataBuilder.TEST_OFFENCE_ID);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyString());
-    }
-
-    @Test
-    void givenAInvalidParameter_whenGetWqResultCodeByCaseIdAndAsnSeqIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.getWqResultCodeByCaseIdAndAsnSeq(TestModelDataBuilder.TEST_CASE_ID, TestModelDataBuilder.TEST_OFFENCE_ID);
-        verify(maatAPIClient).get(any(), anyString(), anyMap(), anyInt(), anyString());
-    }
-
-    @Test
-    void givenAValidParameter_whenFetchResultCodesForCCImprisonmentIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.fetchResultCodesForCCImprisonment();
-        verify(maatAPIClient).get(any(), anyString(), anyMap());
-    }
-
-    @Test
-    void givenAValidParameter_whenFindByCjsResultCodeInIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.findByCjsResultCodeIn();
-        verify(maatAPIClient).get(any(), anyString(), anyMap());
-    }
-
-    @Test
-    void givenAValidParameter_whenGetRepOrderIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.getRepOrder(TestModelDataBuilder.TEST_REP_ID);
-        verify(maatAPIClient).get(any(), any(), any());
-    }
-
-    @Test
-    void givenAValidParameter_whenUpdateCrownCourtOutcomeIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.updateCrownCourtOutcome(UpdateCCOutcome.builder().build());
-        verify(maatAPIClient).get(any(), any(), any());
-    }
-
-    @Test
-    void givenAValidParameter_whenInvokeUpdateAppealSentenceOrderDateIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.invokeUpdateAppealSentenceOrderDate(UpdateSentenceOrder.builder().build());
-        verify(maatAPIClient).get(any(), any(), any());
-    }
-
-    @Test
-    void givenAValidParameter_whenInvokeUpdateSentenceOrderDateIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.invokeUpdateSentenceOrderDate(UpdateSentenceOrder.builder().build());
-        verify(maatAPIClient).get(any(), any(), any());
-    }
-
-    @Test
-    void givenAValidParameter_whenMaatRecordLockedIsInvoked_thenResponseIsReturned() {
-        maatCourtDataService.isMaatRecordLocked(TestModelDataBuilder.TEST_REP_ID);
-        verify(maatAPIClient).get(any(), any(), any());
     }
 
     @Test
